Reduce json validation to an Array.isArray() call
In the previous commit (265f959e906cafcf63aaf180eb9bf02c0657c1c6) a rather
convoluted method was used to attempt to clean and validate the user input
without calling any methods on the json objects, working under the still
potentially true (but almost certainly wildly unlikely) assumption that the
object returned by json may not necessarily be safe. However, accessing
properties (which we certainly have to do here) on an unknown object is just
as dangerous as calling functions, so really the method used in the previous
commit was both inefficient and ineffective.

So, accepting that we are going to call methods on the json object (and,
really, if we can't trust json, what can we trust), the user-inputted json is
now validated as an array with a single Array.isArray() call.